lastquiz-ethical-hacking/backend/requirements_v1.txt
aiofiles==0.7.0
python-multipart==0.0.5
openpyxl==3.0.9
python-magic==0.4.24
pymongo==4.1.1
python-jose==3.2.0
passlib==1.7.4
bcrypt==3.2.0
